+++ crates/forge-runner/src/running.rs
@@ -4,6 +4,7 @@ use crate::gas::calculate_used_gas;
 use crate::package_tests::with_config_resolved::{ResolvedForkConfig, TestCaseWithResolvedConfig};
 use crate::test_case_summary::{Single, TestCaseSummary};
 use anyhow::{Result, ensure};
+use blockifier::execution::contract_class::TrackedResource;
 use blockifier::execution::entry_point::EntryPointExecutionContext;
 use blockifier::state::cached_state::CachedState;
 use cairo_lang_runner::{Arg, RunResult, SierraCasmRunner};
@@ -31,8 +32,8 @@ use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::Use
 use cheatnet::runtime_extensions::cheatable_starknet_runtime_extension::CheatableStarknetRuntimeExtension;
 use cheatnet::runtime_extensions::forge_runtime_extension::contracts_data::ContractsData;
 use cheatnet::runtime_extensions::forge_runtime_extension::{
-    ForgeExtension, ForgeRuntime, add_vm_execution_resources_to_top_call, get_all_used_resources,
-    update_top_call_execution_resources, update_top_call_l1_resources, update_top_call_vm_trace,
+    ForgeExtension, ForgeRuntime, add_resources_to_top_call, get_all_used_resources,
+    update_top_call_l1_resources, update_top_call_resources, update_top_call_vm_trace,
 };
 use cheatnet::state::{
     BlockInfoReader, CallTrace, CheatnetState, EncounteredError, ExtendedStateReader,
@@ -181,7 +182,23 @@ pub fn run_test_case(
     let block_info = state_reader.get_block_info()?;
     let chain_id = state_reader.get_chain_id()?;
 
-    let mut context = build_context(&block_info, chain_id);
+    let mut context = build_context(
+        &block_info,
+        chain_id,
+        &TrackedResource::from(runtime_config.tracked_resource),
+    );
+    let versioned_constants = context
+        .tx_context
+        .block_context
+        .versioned_constants()
+        .clone();
+
+    #[expect(clippy::clone_on_copy)]
+    let tracked_resource = context
+        .tracked_resource_stack
+        .last()
+        .expect("Unexpected empty tracked resource.")
+        .clone();
 
     if let Some(max_n_steps) = runtime_config.max_n_steps {
         set_max_steps(&mut context, max_n_steps);
@@ -240,9 +257,11 @@ pub fn run_test_case(
                     .get_execution_resources()
                     .expect("Execution resources missing")
                     .filter_unused_builtins();
-                add_vm_execution_resources_to_top_call(
+                add_resources_to_top_call(
                     &mut forge_runtime,
                     &vm_resources_without_inner_calls,
+                    &tracked_resource,
+                    &versioned_constants,
                 );
 
                 let ap = runner.relocated_trace.as_ref().unwrap().last().unwrap().ap;
@@ -281,7 +300,7 @@ pub fn run_test_case(
 
     let call_trace_ref = get_call_trace_ref(&mut forge_runtime);
 
-    update_top_call_execution_resources(&mut forge_runtime);
+    update_top_call_resources(&mut forge_runtime, &tracked_resource);
     update_top_call_l1_resources(&mut forge_runtime);
 
     let fuzzer_args = forge_runtime
@@ -293,7 +312,8 @@ pub fn run_test_case(
         .clone();
 
     let transaction_context = get_context(&forge_runtime).tx_context.clone();
-    let used_resources = get_all_used_resources(forge_runtime, &transaction_context);
+    let used_resources =
+        get_all_used_resources(forge_runtime, &transaction_context, tracked_resource);
     let gas = calculate_used_gas(
         &transaction_context,
         &mut cached_state,