+++ crates/forge/src/lib.rs
@@ -4,6 +4,7 @@ use anyhow::anyhow;
 use camino::Utf8PathBuf;
 use clap::{CommandFactory, Parser, Subcommand, ValueEnum};
 use forge_runner::CACHE_DIR;
+use forge_runner::forge_config::ForgeTrackedResource;
 use run_tests::workspace::run_for_workspace;
 use scarb_api::{ScarbCommand, metadata::MetadataCommandExt};
 use scarb_ui::args::{FeaturesSpec, PackagesFilter};
@@ -197,6 +198,10 @@ pub struct TestArgs {
     /// Build contracts separately in the scarb starknet contract target
     #[arg(long)]
     no_optimization: bool,
+
+    /// Specify tracked resource type
+    #[arg(long, value_enum, default_value_t)]
+    tracked_resource: ForgeTrackedResource,
 
     /// Additional arguments for cairo-coverage or cairo-profiler
     #[clap(last = true)]