+++ crates/forge/test_utils/src/running_tests.rs
@@ -10,7 +10,7 @@ use forge::{
 };
 use forge_runner::CACHE_DIR;
 use forge_runner::forge_config::{
-    ExecutionDataToSave, ForgeConfig, OutputConfig, TestRunnerConfig,
+    ExecutionDataToSave, ForgeConfig, ForgeTrackedResource, OutputConfig, TestRunnerConfig,
 };
 use forge_runner::test_target_summary::TestTargetSummary;
 use scarb_api::{ScarbCommand, metadata::MetadataCommandExt};
@@ -66,6 +66,7 @@ pub fn run_test_case(test: &TestCase) -> Vec<TestTargetSummary> {
                         .unwrap()
                         .join(CACHE_DIR),
                     contracts_data: ContractsData::try_from(test.contracts().unwrap()).unwrap(),
+                    tracked_resource: ForgeTrackedResource::CairoSteps,
                     environment_variables: test.env().clone(),
                 }),
                 output_config: Arc::new(OutputConfig {